+++ src/Practice5_MySerializer/IgnoreAttribute.cs
@@ -1,11 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Practice5_MySerializer
+﻿namespace Practice5_MySerializer
 {
     [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field)]
-    public class IgnoreAttribute : Attribute {}
+    public class IgnoreAttribute : Attribute { }
 }
