+++ src/Practice5_MySerializer/JsonNameAttribute.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Practice5_MySerializer
+﻿namespace Practice5_MySerializer
 {
     [AttributeUsage(AttributeTargets.Property | AttributeTargets.Field)]
     public class JsonNameAttribute : Attribute
@@ -13,7 +7,7 @@ namespace Practice5_MySerializer
 
         public string Name => name;
 
-        public JsonNameAttribute(string name) 
+        public JsonNameAttribute(string name)
         {
             this.name = name;
         }
